+++ example/lib/main.dart
@@ -153,6 +153,9 @@ class _HomePageState extends State<HomePage> {
                     });
                   },
                 ),
+                pieceOrientationBehavior: playMode == Mode.freePlay
+                    ? PieceOrientationBehavior.opponentUpsideDown
+                    : PieceOrientationBehavior.default_,
                 pieceShiftMethod: pieceShiftMethod,
               ),
               state: ChessboardState(
@@ -163,7 +166,6 @@ class _HomePageState extends State<HomePage> {
                         : InteractableSide.black),
                 validMoves: validMoves,
                 orientation: orientation,
-                opponentsPiecesUpsideDown: playMode == Mode.freePlay,
                 fen: fen,
                 lastMove: lastMove,
                 sideToMove: